ref(ngrx-ducks): extract test data and action decorator into own files
