+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/index.ts
@@ -1,0 +1,4 @@
+export * from './with-empty-action-type';
+export * from './with-null-action-type';
+export * from './with-undefined-action-type';
+export * from './with-valid-action-type';
