+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-empty-action-type.ts
@@ -1,0 +1,8 @@
+import { Action } from '../../action.decorator';
+
+export class WithEmptyActionType {
+  @Action('')
+  emptyIsNotValid(state: number, payload: number) {
+    return state + payload;
+  }
+}
