+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-null-action-type.ts
@@ -1,0 +1,8 @@
+import { Action } from '../../action.decorator';
+
+export class WithNullActionType {
+  @Action(null as any)
+  nullIsNotValid(state: number, payload: number) {
+    return state + payload;
+  }
+}
