+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-undefined-action-type.ts
@@ -1,0 +1,8 @@
+import { Action } from '../../action.decorator';
+
+export class WithUndefinedActionType {
+  @Action(undefined as any)
+  undefinedIsNotValid(state: number, payload: number) {
+    return state + payload;
+  }
+}
