+++ ngrx/ducks/src/lib/core/wired-actions/mocks/duck-candidates/with-valid-action-type.ts
@@ -1,0 +1,10 @@
+import { Action } from '../../action.decorator';
+
+export const actionType = '[Counter] Set Number';
+
+export class WithValidActionType {
+  @Action(actionType)
+  increase(state: number, payload: number) {
+    return state + payload;
+  }
+}
